pyproject.toml
[project]
name = "poc-dj-entities"
version = "0.0.1"
description = "Default template for PDM package"
authors = [
    {name = "gcca", email = "_@gcca.dev"},
]
dependencies = ["Django>=6.0.2", "djangorestframework>=3.16.1", "drf-spectacular>=0.29.0", "djangorestframework-simplejwt>=5.5.1"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "AGPL-3.0-or-later"}


[tool.pdm]
distribution = false

[dependency-groups]
dev = [
    "django-stubs>=5.2.9",
    "pre-commit>=4.5.1",
]
